adding updated/separate CLI code
I think something went wrong back in late sept.. I put version 0.32 up and then it was reverted back to 0.30 possibly by accident. Here is the latest but uploading separately as there have been several changes that I am afraid to interfere with. 

Changelog since 0.30:

# ver 0.33 
# - update archive_data function to be MUCH faster when data is highly segmented
# - some typos, bugs fixed
# - force stations and NSLC to be uppercase

# ver 0.32 22/09/2024
# - finish adding auth capability, change format in config files
# - add "catalog" option to [EVENT] settings to use a pre-downloaded "QuakeML" xml catalog instead searching for and downloadin one
# - remove .lower() to process config items as this is done by default
# - allow some config settings to be case sensitive (e.g. AUTH)
# - add "dist_deg" kwarg to get_p_s_times instead of calculating it twice

# ver 0.31 06/09/2024
# - select only highest samplerate channels for event downloads

the main changes are the new version of archive_request which is 100x faster and the addition of the option to use authenticated clients for restricted data